Mod: move service name mapping
